Relocate class name utility and update imports
