+++ components/ui/3d-card.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/utils";
 
 import React, {
   createContext,
