+++ components/ui/TextGenerateEffect.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useEffect, useMemo } from "react";
 import { motion, stagger, useAnimate } from "framer-motion";
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/utils";
 
 export const TextGenerateEffect = ({
   words,
